requirements.txt
Django==2.2.4
django-phone-field==1.8.1
pytz==2020.1
sqlparse==0.3.1
